phase 5 skeleton: state machine, script VM, dialogs-lite
- GameContext state machine {Loading=7, Playing=3, Dying=13} with legacy
  setState semantics; revived GameLoop as fixed-step driver (15ms quanta,
  dt clamp 125ms) wired through AppContext::run(); Main.cpp reduced to
  init + run
- ScriptVM: faithful 20-thread pool (state/unpauseTime/-1 external
  resume), big-endian operands, ~40-opcode subset incl. EVAL/JUMP/
  CALL_FUNC/ITEM_COUNT/DOOROP/EVENTOP/GIVEITEM/WAIT/ABORT/DIALOG-lite/
  ENTITY_FRAME; tile triggers (leave-before-move, FACE+enter on arrival,
  spawn-tile event, staticFuncs 0/6)
- scripted doors: setLineLocked port (tileNum bit flip + def re-lookup),
  EV_DOOROP n2 semantics (quiet-bit snap), DoorAnim::ownerThread resume
- dialog-lite: gray HUD panel, E-dismissal stepping, chained dialogs;
  text pipeline fixes: kTextMap loading, legacy glyph clamp (space
  rendered as garbage underscore), word wrap, dehyphenate parity
- lazy sprite-texture creation (ensureSpriteTexture) fixing post-unlock
  door vanish; angle-wrap turn freeze fix (raw accumulated angles per
  legacy); media reference-record resolution follow-up
- docs: game-flow + tile-events VM research, phase5 spec, ADR-0003,
  journal/status/PLAN updates

+++ new_src/text/Font.cpp
@@ -82,7 +82,10 @@ void Font::drawChar(SpriteBatch& batch, char c, int x, int y, int rotateMode,
 
 	int index1, index2;
 	getCharIndices(c, &index1, &index2);
-	if (index1 > 143 || index2 < 0 || index2 > 143) {
+	// Out-of-range guard from src/Graphics.cpp:641-652. index1 must compare
+	// unsigned so negative results (space 0x20 -> -1) also take the '?' cell
+	// instead of sampling outside the sheet.
+	if ((uint32_t)index1 > 143u || index2 < 0 || index2 > 143) {
 		index1 = 30; // '?' fallback
 		index2 = 0;
 	}
